test: refactor image imports and add tests for SpectralLibImage functionality
